Fixes an error where input file path wasn't taken as absolute.
When input file was not in the current directory, assembler would fail printing
a pointless error message. This was caused by the fact that when new buffer
was pushed it used the same mechanism that is used for file inclusion. This
is incorrect and the input file should be always taken as an absolute path.