Prevent DS directive from emitting empty data.
DS directive was emitting empty (zero) data into a section.
This is not how qasm used to work and tends to produce
gigantic files. This change makes it to skip generating
data for all DS statements. This also makes the programmer
responsible for initialising their variables defined with DS
directives.